기억저장소 구현중

+++ Assets/Scripts/Data/InGameData.cs
@@ -3,7 +3,7 @@ public class InGameData
     public int AccumulatedGold;
     public int Gold { get; set; }
     public int Gem { get; set; }
-    public float Experience { get; set; } = 0;
+    public int Experience { get; set; } = 0;
 
-    public float NextEXP { get; set; } = 50;
+    public int NextEXP { get; set; } = 50;
 }
